Rename otap_batch_processor to batch_processor
