+++ rust/otap-dataflow/crates/otap/src/lib.rs
@@ -22,7 +22,7 @@ pub mod otlp_receiver;
 pub mod otlp_exporter;
 
 // OTAP batch processor
-pub mod otap_batch_processor;
+pub mod batch_processor;
 
 // Retry processor that is aware of the OTAP PData/context.
 pub mod retry_processor;
